text-to-speech/requirements.txt
ffmpeg-python == 0.2.0
future == 0.18.3
gTTS == 2.3.1
